chore: update footer and add failing slug build
