+++ components/Footer/index.jsx
@@ -197,8 +197,8 @@ export default function Footer({ noCta }) {
                       </Link>
                     </div>
                     <div className={s.groupLink}>
-                      <Link href="/partners">
-                        <a>Solution partners</a>
+                      <Link href="/tech-partners">
+                        <a>Technology partners</a>
                       </Link>
                     </div>
                     <div className={s.groupLink}>
